Refactor: Move models into their own folder/files
